+++ changelog.md
@@ -52,6 +52,7 @@ Also EnableSolanaAddressLookupTable feature flag should be set.
 * [4182](https://github.com/zeta-chain/node/pull/4182) - observe Sui inbound from previous gateway package after authenticated call upgrade
 * [4254](https://github.com/zeta-chain/node/pull/4254) - add additional support for zetaclient public DNS name
 * [4342](https://github.com/zeta-chain/node/pull/4342) - add metrics for monitoring inbound voting through blockscan and trackers
+* [4377](https://github.com/zeta-chain/node/pull/4377) - add number of connected peers to tss keygen log
 
 ### Tests